+++ Model/Cache.php
@@ -8,9 +8,15 @@ declare(strict_types=1);
 
 namespace Goomento\PageBuilder\Model;
 
+use Exception;
 use Magento\Framework\App\Cache\StateInterface;
 use Magento\Framework\App\CacheInterface;
 use Goomento\PageBuilder\Model\Cache\Type\PageBuilder;
+use Magento\Framework\App\Cache\TypeListInterface;
+use Magento\PageCache\Model\Cache\Type;
+use Magento\Framework\App\Cache\Type\Block;
+use Zend_Cache;
+use Zend_Json;
 
 class Cache
 {
@@ -27,17 +33,24 @@ class Cache
      * @var bool
      */
     private $enabled;
+    /**
+     * @var TypeListInterface
+     */
+    private $typeList;
 
     /**
      * @param CacheInterface $cache
      * @param StateInterface $cacheState
+     * @param TypeListInterface $typeList
      */
     public function __construct(
         CacheInterface $cache,
-        StateInterface $cacheState
+        StateInterface $cacheState,
+        TypeListInterface $typeList
     )
     {
         $this->cache = $cache;
+        $this->typeList = $typeList;
         $this->enabled = $cacheState->isEnabled(PageBuilder::TYPE_IDENTIFIER);
     }
 
@@ -47,6 +60,22 @@ class Cache
     public function isEnabled()
     {
         return (bool) $this->enabled;
+    }
+
+    /**
+     * Set invalidate the cache
+     *
+     * @return void
+     */
+    public function invalid()
+    {
+        if ($this->isEnabled()) {
+            $this->typeList->invalidate([
+                PageBuilder::TYPE_IDENTIFIER,
+                Type::TYPE_IDENTIFIER,
+                Block::TYPE_IDENTIFIER
+            ]);
+        }
     }
 
     /**
@@ -111,7 +140,7 @@ class Cache
             return null;
         }
 
-        return $this->cache->clean(\Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG, (array) $tags);
+        return $this->cache->clean(Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG, (array) $tags);
     }
 
     /**
@@ -122,8 +151,8 @@ class Cache
     {
         if (!is_scalar($data)) {
             try {
-                $data = \Zend_Json::encode($data);
-            } catch (\Exception $e) {}
+                $data = Zend_Json::encode($data);
+            } catch (Exception $e) {}
         }
 
         return $data;
@@ -137,8 +166,8 @@ class Cache
     {
         if (is_string($data)) {
             try {
-                $data = \Zend_Json::decode($data);
-            } catch (\Exception $e) {}
+                $data = Zend_Json::decode($data);
+            } catch (Exception $e) {}
         }
 
         return $data;